Add theme prop and capitalize theme title
Refactor the Body component to accept a theme prop and capitalize the first letter of the theme title in the heading. The theme prop is used to determine the theme of the heading, which is displayed in the format "This is {theme} Theme".

This change is intended primarily for testing purposes, but it also improves the modularity and reusability of the Body component by allowing it to be used with different themes.