Refactor:  buttons to use dynamic attributes with createDynamicButton utility function
This commit refactors the existing button styles to use a createDynamicButton utility function that applies dynamic attributes to the component based on the presence of an href prop. This reduces code duplication and makes the styles more consistent across different button types.

A new utils folder has been added to the project to store the buttonUtils.js file containing the createDynamicButton function.

The updated DefaultBtn, ElevatedBtn, FilledBtn, TonalBtn, and OutlinedBtn components now use the createDynamicButton function to generate the appropriate HTML tag for each button based on the presence of an href prop.

All existing button styles and functionality remain unchanged.