refactor: Allow text or children prop in Button component
This commit refactors the Button component to allow either a text or children prop to be passed to the component for the button label. The component uses an enum object called ButtonConfig to determine which style to apply to the button. The label of the button is determined by checking whether the children prop is passed, and if not, the text prop is used instead. This change provides more flexibility for the developer to choose how to pass the button label to the component, making the Button component more user-friendly.