refactor: Rename rest parameter to props in Button component
This commit renames the rest parameter to props in the Button component. The rest parameter is commonly used to pass down all the remaining props not explicitly destructured. However, in this case, it would be clearer to rename the parameter to props for better readability and to be more consistent with React best practices. This update does not affect the functionality of the component.