refactor: rename Body component to MainBody and update its import
The Body component has been renamed to MainBody to better reflect its purpose. Its import has been updated accordingly in the App component.

BREAKING CHANGE: This change affects the import path of the Body component.

feat: add toggleTheme prop to FloatingActionButton component

The toggleTheme prop has been added to the FloatingActionButton component, which allows users to toggle the theme of the application with a single click.

BREAKING CHANGE: This change introduces a new prop to the FloatingActionButton component.