feat: add DoubleArrowDown component
Add the DoubleArrowDown component, which renders a double arrow icon. This icon consists of two arrows, one pointing up and one pointing down. The arrows are pink in color and have a width and height of 24 pixels. This component can be used to indicate that a list can be sorted both ways.