feat: add Edge launch configuration for debugging
Add a launch configuration to launch Microsoft Edge when debugging the React app on localhost. This configuration sets the "type" to "msedge", "request" to "launch", and "name" to "Launch Edge against localhost". The "url" property is set to "http://localhost:3000", which is the default port used by the React development server. The "webRoot" property is set to "${workspaceFolder}", which is the root directory of the project. This configuration allows developers to easily debug the React app using Microsoft Edge as the browser.