feat: Implement routing in main application component
The main App component has been updated to use the `react-router-dom`
library. This library allows us to set up routing for our application,
providing a way to define multiple routes and navigate between
different components without having to reload the entire page.

A `Router` component wraps the whole application, and a `Routes`
component now contains two `Route` components. The first route is
associated with the `MainBody` component and maps to the root ("/")
URL. The second route maps to the "/sandbox" URL and is associated
with a new `Sandbox` component for testing.