feat: Add LightBulbIcon component and use it in JSX
This commit adds a new LightBulbIcon component to the project, which renders a light bulb icon using SVG. The component is added to the project's icon library and can be used throughout the application.

In addition to adding the LightBulbIcon component, this commit updates several components to use the new icon in JSX. For example, the `Header` component now uses the LightBulbIcon component to render a light bulb icon in the header.

Overall, this commit adds a new icon to the project and improves the consistency and usability of the application's icon library.