feat: Add mediaQueries.js file for managing media queries and padding sizes
This commit adds a new `mediaQueries.js` file to the project, which provides a centralized location for defining media queries and padding sizes that can be used throughout the application. The file defines a `media` object that contains media query functions for each breakpoint defined in the `BREAKPOINTS` object, as well as a `padding` property that contains padding sizes defined in the `PADDING_SIZES` object.

The `media` object is created using the `reduce()` method to iterate over the keys in the `BREAKPOINTS` object and create a property for each breakpoint in the `media` object. The `padding` property is created using a similar approach, using the `reduce()` method to iterate over the keys in the `PADDING_SIZES` object and create a property for each padding size in the `media.padding` object.