refactor: update Nav component styles
This commit refactors the styles of the Nav component. The changes involve
removing the media query and updating the class name from ".svg-container" to
".logo".

In the Nav component's styled-component definition, the media query block has
been removed as it was no longer necessary. Additionally, the class name has
been updated to ".logo" to provide a more descriptive and meaningful name for
the element.