refactor: Enhance Contact component and extract initial form state
Refactored the Contact component to integrate with the updated
useValidation hook and the newly added submitContactForm function. This
change aligns the component with the revised form validation and
submission process, improving modularity and maintainability. The form
submission logic now includes serverless functionality through Netlify
and SendGrid integration.

- Extracted the initialState for the contact form into a separate file,
  `contactFormInitialStates.js`, enhancing readability and reusability.
- Simplified the `useValidation` hook usage by removing the initialState
  parameter, now directly importing `contactFormInitialState`.
- Added new state variables `isSubmitting` and `submissionStatus` to
  manage form submission state and user feedback.
- Implemented `resetForm` function for resetting the form after
  submission.
- Refined form submission process with `handleSubmit`, integrating
  serverless email functionality.
- Updated `handleEnterKey` for improved field navigation.
- Enhanced UI feedback with dynamic `buttonText` during form submission.

This refactor enhances the overall structure of the Contact component,
making it more efficient and user-friendly, while also ensuring better
separation of concerns in line with React's component-based design.